fixed URI naming, user-jogging relation, crud methods for joggings

+++ src/main/java/com/example/yellow/model/JoggingModel.java
@@ -14,8 +14,8 @@ import java.time.LocalDateTime;
 public class JoggingModel {
 
     private Long id;
-    private int distance;
-    private int time;
+    private Long distance;
+    private Long time;
     private LocalDateTime dateTime;
 
 }
